update columns from float to decimal

+++ db/migrate/20140411011049_create_procedure.rb
@@ -3,7 +3,7 @@ class CreateProcedure < ActiveRecord::Migration
     create_table :procedures do |t|
       t.string :code, null: false
       t.string :description, null: false
-      t.float :work_rvu, null: false
+      t.decimal :work_rvu, null: false
 
     end
   end
